+++ src/osha/oira/__init__.py
@@ -1,5 +1,6 @@
 from zope.i18nmessageid import MessageFactory
 from logging import getLogger
+import patch_passwordreset
 
 _ = MessageFactory("euphorie")
 log = getLogger(__name__)