+++ src/osha/oira/content/login.py
@@ -1,7 +1,12 @@
 from five import grok
-from plonetheme.nuplone.skin import login
-from zope.interface import Interface
 from osha.oira.interfaces import IOSHAContentSkinLayer
+from plonetheme.nuplone.skin import login
+from plonetheme.nuplone.skin import pwreminder
+from Products.CMFCore.interfaces import ISiteRoot
+from zope.i18nmessageid import MessageFactory
+from zope.interface import Interface
+
+MF = MessageFactory("nuplone")
 
 grok.templatedir("templates")
 
@@ -13,3 +18,34 @@ class Login(login.Login):
     grok.layer(IOSHAContentSkinLayer)
     grok.name("login")
     grok.template("login")
+
+
+class OSHARequestPasswordForm(pwreminder.RequestPasswordForm):
+    """ Override so that we can change some labels
+    """
+    grok.context(ISiteRoot)
+    grok.layer(IOSHAContentSkinLayer)
+    grok.name("request-password-reset")
+
+    def updateFields(self):
+        super(OSHARequestPasswordForm, self).updateFields()
+        self.fields["login"].field.title = MF(
+            u"label_email", default=u"E-mail address")
+
+
+# XXX
+# The following leads to a Configuration Conflict Error on startup
+# Therefore the required piece of code is now located in
+# patch_passwordreset.py
+
+# class OSHAPasswordReset(pwreminder.PasswordReset):
+#     """ Override so that we can change some labels
+#     """
+#     grok.context(ISiteRoot)
+#     grok.layer(IOSHAContentSkinLayer)
+#     grok.name("reset-password")
+
+#     def updateFields(self):
+#         super(OSHAPasswordReset, self).updateFields()
+#         self.fields["login"].field.title = MF(
+#             u"label_email", default=u"E-mail address")