+++ src/osha/oira/patch_passwordreset.py
@@ -1,0 +1,12 @@
+from plonetheme.nuplone.skin.pwreminder import PasswordReset
+from zope.i18nmessageid import MessageFactory
+
+MF = MessageFactory("nuplone")
+
+
+def updateFields(self):
+    super(PasswordReset, self).updateFields()
+    self.fields["login"].field.title = MF(
+        u"label_email", default=u"E-mail address")
+
+PasswordReset.updateFields = updateFields